finish merging notebooks
