+++ pgml-dashboard/app/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import *
+
+admin.site.register(Notebook)
+admin.site.register(NotebookCell)
