--- pgml-dashboard/notebooks/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from notebooks.models import *
-
-admin.site.register(Notebook)
-admin.site.register(NotebookCell)
