+++ pgml-dashboard/pgml_dashboard/settings.py
@@ -59,7 +59,6 @@ INSTALLED_APPS = [
     "django.contrib.humanize",
     "rest_framework",
     "django_filters",
-    "notebooks",
     "request",
 ]
